# add an option to enable native compilation optimization