After I enable it, I can observe speedup in some cases. For `examples/water/se_e2_a`, compress training FP32 networks:
```sh
CUDA_VISIBLE_DEVICES= OMP_NUM_THREADS=8 TF_INTRA_OP_PARALLELISM_THREADS=8 TF_INTER_OP_PARALLELISM_THREADS=2 dp train input.json -f frozen_model_compressed.pb
```

Before enabling it,
```
DEEPMD INFO    batch     100 training time 1.67 s, testing time 0.03 s
DEEPMD INFO    batch     200 training time 1.25 s, testing time 0.03 s
DEEPMD INFO    batch     300 training time 1.23 s, testing time 0.03 s
DEEPMD INFO    batch     400 training time 1.24 s, testing time 0.03 s
DEEPMD INFO    batch     500 training time 1.24 s, testing time 0.03 s
DEEPMD INFO    batch     600 training time 1.23 s, testing time 0.03 s
DEEPMD INFO    batch     700 training time 1.24 s, testing time 0.03 s
DEEPMD INFO    batch     800 training time 1.24 s, testing time 0.03 s
DEEPMD INFO    batch     900 training time 1.24 s, testing time 0.03 s
DEEPMD INFO    batch    1000 training time 1.24 s, testing time 0.03 s
```
After enabling it,
```
DEEPMD INFO    batch     100 training time 1.60 s, testing time 0.03 s
DEEPMD INFO    batch     200 training time 1.10 s, testing time 0.03 s
DEEPMD INFO    batch     300 training time 1.10 s, testing time 0.03 s
DEEPMD INFO    batch     400 training time 1.10 s, testing time 0.03 s
DEEPMD INFO    batch     500 training time 1.11 s, testing time 0.03 s
DEEPMD INFO    batch     600 training time 1.11 s, testing time 0.03 s
DEEPMD INFO    batch     700 training time 1.10 s, testing time 0.03 s
DEEPMD INFO    batch     800 training time 1.10 s, testing time 0.03 s
DEEPMD INFO    batch     900 training time 1.10 s, testing time 0.03 s
DEEPMD INFO    batch    1000 training time 1.13 s, testing time 0.03 s
```
About 10% improvement.